JVMCBC-1120: ClusterConfig#allNodeAddresses needs to take global config into account.
Right now the allNodeAddresses method does not take the global config into account,
which can lead to issues if there is no bucket open then the core watchdog will
report (debug level) info that the node count does not add up and trigger a
reconfiguration. it's not harmful, but needs to be patched up anyways.

Change-Id: I18228ea21212cb8465f090bdde79a5eeb8ac95ee
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/177974
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>