JVMCBC-1110: Client is not reliably removing itself from client records on shutdown
Part of the transactions protocol is that each client (application)
writes a small entry into the _txn:client-record of each collection
in the 'cleanup set' (the collections it's actively cleaning up).

During shutdown it removes itself from each such _txn:client-record.

Intermittently, this removal is not happening.  This isn't
particularly crucial as each client writes a heartbeat, and so
another client in future will detect this one has expired, and remove
it.  But it does cause intermittent FIT test failures.

This patchset redoes cleanup shutdown, with a clearer and safer
thread-safety model to avoid what I believe was the cause - that
a client can be about to write itself into the client record, after
such threads have been told to stop, and before it's checked that
stop signal.

Now we set the stop signal, and wait for all threads to complete
before proceeding to remove from the client records.

In addition improving cleanup in this way:

* Fixed some missed logging redactions.
* Improved logging.
* Renamed a method to better reflect what it does.
* Made shutdown (more) reactive to avoid thread starvation.
* Put the cleanup threads on the correct transactions cleanup
  scheduler.

Change-Id: I17fd9328ae568637017cfe2d552ea537219977f8
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178039
Reviewed-by: <matthew.bray@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>