JVMCBC-1121: Should warn after 2 hours of failing to cleanup a transaction, not 2 days
Usually a transaction finishes committing or rolling back
before control is returned to the application.

In some rare cases (application crash etc.) this isn't
possible, and the transaction is left for the async cleanup
algo to finish.

Usually this algo will find the transaction within one
minute (default settings) and finish the commit or rollback.

If for some reason it can't, it will move on, and usually
discover the transaction a minute later, and try again.

It will repeat this forever.

This is an opaque, background process, so to provide some
visibility to the user in case of bugs etc., we start
logging about such failures after a set period.

That period was supposed to be and specced as 2 hours, rather
than the 2 days we are currently using.

Change-Id: I8cbc8db83bf2ce34c7cd18f5783fd8e0600f0e69
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178042
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>