CBD-5023: Merge FIT Java performers
There are currently 3 Java performers:

* transactions-fit-performer/performers/java
  The OG, used for testing the transactions library.
* couchbase-jvm-client/java-fit-performer
  For testing the integrated version of transactions.
* transactions-fit-performer/performers/jvm/*
  Java, Scala and Kotlin performers that only implement
  enough of the GRPC to handle the SDK performance
  and integration testing (not transactions).

This changeset merges the latter two, to reduce our
maintenance burden and (in future) permit performance
testing of transactions.

This includes adding the Scala and Kotlin performers
to this project, along with the CorePerformer
they and the Java performer now share.

It's just moving the existing tested and reviewed
performer code from transactions-fit-performer,
and tweaking the pom.xmls to reflect the structure of
couchbase-jvm-clients.

Change-Id: I13c301e76714ff05ed097e4aba8c5542b52f4f4e
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178405
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>