JCBC-1987 Fix buildDeferredIndexes on default collection
Motivation
----------
`QueryIndexManager.buildDeferredIndexes` on default collection
fails if deferred indexes with different name exist in other collection.

Modifications
--------------
When calling getAllIndexes to determine which indexes to build,
always narrow the results to a specific collection.

Update the Javadoc to clarify that `buildDeferredIndexes` always
targets a single collection. (This is different from `getAllIndexes`,
which can return all indexes in a bucket, scope, or collection
depending on which options are specified.)

Add a regression test.

Change-Id: Id1d47c698a52d8357af8a93a67ced21db02b2142
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178429
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>