JCBC-1989: Enable try-with-resources for Cluster
If there is a trivial use case where only a handful of operations
need to be performed and the SDK does not need to be connected
for a longer period of time the try-with-resources pattern comes
in handy.

By implementing Closeable we can support snippets like this:

try (Cluster cluster = Cluster.connect("127.0.0.1", "user", "password")) {
  Bucket bucket = cluster.bucket("travel-sample");
  bucket.waitUntilReady(Duration.ofSeconds(10));
  // ...
}

This will automatically call cluster.disconnect() and free all resources
and connections.

Change-Id: Ia58548091c139fd14d0fc8a327cc23aa2b3a865b
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178538
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>