JVMCBC-1125: CompletionException intermittently seen from transactions
Intermittently we see a CompletionException in
Hooks.onErrorDropped, when the cleanup window is set
very low (much lower than a user likely would).

The cleanup window is the time it takes us to poll
all 1,024 ATRs.  The default is 60 seconds which gives
58ms per ATR.  That's generally plenty for them to
happen in serial.

But, we do allow the operations to overlap (by using a
flatMap), so if the cleanup window is small enough
and/or handling an ATR takes longer than expected,
concurrent ATR operations can happen.

When one of those operations hits an error
(probably DocumentNotFoundException, which is an expected
one as the ATR often won't exist), because of the way
reactor works, all concurrent operations are cancelled.

But one of those concurrent operations may hit an error
(such as DocumentNotFoundException) immediately after
being cancelled.  Reactor has no way to raise that error
properly now, so it raises it on Hooks.onErrorDropped.

I'm solving this with a MonoBridge.  This effectively
allows concurrent operations, but a cancel/error signal on
one does not automatically cancel the rest.  Instead the
concurrent operations are allowed to cleanly finish, but
their onNext signals are blocked from propagating.

Change-Id: I652ae9b43761c02995076642754efffbfdad574c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/178785
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>