JCBC-1993, SCBC-367: NPE when no search index exists on the cluster.
Both scala and java are not correctly handling a server response where
no index at all is defined on the cluster, which returns
'{ "status" : "ok", "indexDefs" : null }'.

This change makes sure that if null is returned, an empty list/seq
is bubbled up to the caller instead of returning null.

Change-Id: If4cf04c9390af04c875719dd6fe1dba5dcf70197
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/179004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>