JVMCBC-1137: Add initial support for returning write and read units
Some builds of the server have the concept of KV operations
costing read and write units.  These can be returned if the
HELO::ReportUnitUsage flag is negotiated.

This changeset negotiates that flag and rtrieves the reported
read and write units for a subset of operations.

The initial use-case for this is to expose this information
into the transaction logs to assist with MB-53127, so
the only operations it is exposed in currently are those
used by transactions.

The set of operations might be extended further in future
if more use-cases are required - for example, to expose
fully in the SDK *Result objects.

The units are only exposed in the *Response objects,
and the key-value error context, since transactions
uses those directly.

It's a moderately impactful change, but we are now well
placed to extend this support to other *Response objects
in the future, plus to easily support any new
flexible framing extras.

Change-Id: I6b232bd35589352d64e8d2036aa53bc88a918e8c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/179070
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>