CBD-5086: Pathfind transactions tracing and metrics - adding metrics
Adding support for some transactions metrics:
* Total transaction count.
* Total attempt count.

The metrics should be regarded as volatile and may change
(names, attributes etc.) as the design process continues.

There's now a reasonable amount of Core-level state for
transactions, so I've created a CoreTransactionsContext to
hold it going forwards.  I didn't move the existing state
into it to reduce the scope of this change.

Change-Id: I39549272d29b002f97b4bed355f7410e103a8973
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/179174
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>