JVMCBC-1143: Investigate failing the top span with an error if completed with exception - StatusCode.ERROR
Reversing the (unpublished) decision to have span.recordException
also set StatusCode.ERROR.  I've found in testing that it can be
useful to be able to record an exception on a span without marking
it as failed, and in fact this can be a common occurrence.  For instance,
when we retry an operation (in the same span) on a transient failure.

Also marking this new method Volatile for safety.

Change-Id: I1a386a1706ce9600e8bbb91337d14955ac1f22cd
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/179729
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>