JCBC-2002: expiryTime should always return empty optional if no expiry set.
Motivation
----------
At the moment expiryTime will return three different values depending on
the situation:

 1) None if withExpiry is not set in the options.
 2) Some(epoch) if withExpiry is true but no expiry on the document
 3) Some(actual expiry) if withExpiry is true and has an expiry set

Case #2 does not align with user expecations and should return None
as well (since None == no expiry set).

Modification
------------
Adjust the code and tests to return None even if withExpiry is set
to true but the document does not have an expiry value set.

Result
------
Returns None if withExpiry is not set or when it is set and the
document does not have an expiry.

Change-Id: Iae5561aeebe6032d5359a60b3d9392aa8b67efb7
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/180300
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: David Nault <david.nault@couchbase.com>