JVMCBC-1157: Do not mistake magma buckets for ephemeral buckets.
Motivation
----------
With Couchbase Server 7.1 and later, the magma storage type for
the Couchbase bucket got introduced. Unfortunately the heuristics
on how we decide for a ephemeral bucket now overlap with magma
storage types.

This is because previously we had to determine if a bucket is
an ephemeral bucket based on the missing COUCHAPI feature, but
this is also missing for magma buckets. This means that now
operations like PersistTo (which check for ephemeral) do not
work anymore even though they should.

Modifications
-------------
Thankfully in server 7.1 a new bucketType string in the terse
config got introduced, which we can now use to properly determine
if it's an ephemeral bucket or a couchbase bucket with the magma
storage type.

NOTE: our configs have a type() which still returns COUCHBASE as
a BucketType, even for ephemeral buckets. This is done for backwards
compatibility since it is not clear which external code may use it -
but we can still utilize it for our ephemeral() calculation.

Result
------
Allow PersistTo operations against magma storage types and not
pretend they are ephemeral.

Change-Id: Ibfb7331fafb3c17e2f13751a9a48be248a450989
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/181731
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>