JVMCBC-1159: Handle query_context changes for transactions
Some cluster versions now require sending a query_context
on all queries.

Handle this in transactions by using the first scoped query
(e.g. `ctx.query(scope, "...")`, and using that scope for
the BEGIN WORK, COMMIT and ROLLBACK statements.

Without that, or if the user performs an unscoped query,
they will get back a 1197 tenant missing error from query.
So added some handling to fail nicely with a
FeatureNotAvailableException cause.

Testing is in FIT, as usual.

Change-Id: I59d10e316a071b77c4657c871842e1e91d42569d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182038
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>