Gardening: Add CbCollections.listCopyOf, setCopyOf, and mapCopyOf
These are backports of Java 10's List.copyOf, Set.copyOf,
and Map.copyOf.

They differ from the existing copyToUnmodifiableList/Set in that
the new methods throw NPE if the argument is null or contains null.
This is useful for making defensive copies in cases where
passing null (or null element/keys/values) is an error.

Also, annotate the nullable parameters of other CbCollections
methods as @Nullable.

Change-Id: Ib317b01cbae542e6ba5c15740bc8148d76b764e0
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/181912
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>