JCBC-1958: Bucket creation or update is failing against 7.X CE
When we create a bucket, we don't send compressionMode if it's
been left at default (passive), as CE will reject any request
with that set.

However, when getting a bucket, though compressionMode is
not returned by a CE server, we set CompressionMode=Off.  So
when updateBucket() is called, we try to send compressionMode,
and the server rejects with an error "Compression mode is
supported in enterprise edition only".

Fixing by passing through that the server did not return
compressionMode, in the form of nul.

Change-Id: I5fefccce5ec6f1c42038f6cc6fec77b779532549
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182095
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>