JCBC-2013 Make ScanTerm immutable
Motivation
----------
Prevent a user from modifying the static byte arrays holding
the minimum and maximum terms.

Immutable types are nice.

Modifications
-------------
Remove the public static byte array "constants", since
there's no way to prevent a user from modifying elements
of a visible array.

Make a defensive copy of the array in the constructor, and again
in the id() method.

Now that ScanTerm is fully immutable, share the "minimum"
and "maximum" ScanTerm instances instead of the byte arrays.

Change-Id: Id5a1504b934ce305a1f2abb51a1b779b82d6249c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182346
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>