Tests - disabling cleansIdleConnections
Tracked in JVMCBC-1023.  Other tests in this file
are disabled already for the same reason - a
mocking-related error that does not make sense,
and only happens intermittently.  Some time
has been spent already on trying to resolve,
unsuccessfully.  In the interests of getting to green,
temporarily disabling.

See JVMCBC-1023 for full error output and stacktrace.

Change-Id: Id3386ab20a1f0dbd56264165692ba529b6443646
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182402
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>