JVMCBC-1166 Deprecate NodeConnectedEvent in favor of NodeCreatedEvent
Motivation
----------
Despite its name, NodeConnectedEvent is fired prior to
connecting to services on the node. This is confusing for users.

Modifications
-------------
Deprecate NodeConnectedEvent. Change its severity to VERBOSE so it's
less likely to appear in the log.

Add NodeCreatedEvent (the name parallels CoreCreatedEvent).
Publish it at the same time as NodeConnectedEvent.

Continue publishing NodeConnectedEvent in case a user expects it.

Change-Id: I2bc75924f604e89e349049b74ffcb3342945b312
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182648
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>