SCBC-380: Bucket creation or update is failing against 7.X CE
Bucket creation and update is failing against CE
with errors related to maxTTL and compression.

Like JCBC-1958, this requires BucketSettings to be
able to pass through that the server did not return
these fields.

Unfortunately, while with Java we can represent
that with nullability, with Scala nullability is
very non-standard, and Option/Try/Either is
preferred.

So, a breaking change is made to the BucketSettings
API, which is marked @Volatile.

Added tests for compression mode and maxTTL on CE.
These failed due to some server error messages not
being handled, so changed over to use the core bucket manager
implementation, which handles them correctly.

All tests continue to pass, now on CE and EE.

Change-Id: Ic2df19a1c5bb1c93d13fbea359dc062641dce37c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/182777
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>