Tests - fix DefaultConfigurationProviderIntegrationTest
retriesOnBucketNotFoundDuringLoadException is experiencing
the race that ConfigWaitHelper is designed to prevent,
and periodically on CI raising SeedNodeOutdatedException
as the cause of the BucketOpenRetriedEvent
rather than the expected BucketNotFoundDuringLoadException.

This is because DefaultConfigurationProvider emits an empty
config, which races with the real config raised by openBucket.
Essentially the logic starts adding services based on the
openBucket config, but then finds the node has been removed
by the empty config, and bails out with
SeedNodeOutdatedException.

This is all hidden from users by retries and causes no problems
there.  It only affects these granular tests that are requiring
a deterministic outcome.  So using ConfigWaitHelper to wait
for the initial empty config.

Change-Id: If60f75280841f40ce288f7441ecda58054bc6b49
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/183225
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>