JCBC-2028 Throw exception instead of ignoring connection string scheme or parameters
Motivation
----------
The connection string scheme and parameters can specify
important security settings that should not be ignored.

Modifications
-------------
If a pre-built cluster environment is incompatible with the
connection string, throw IllegalArgumentException (instead of
ignoring bits of the connection string and logging a warning).

Deprecate `ConnectionStringIgnoredEvent`, since it's no longer
published.

Change-Id: Idd22106b3a924ff318195b3439ead1fd0a33b1f7
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/183520
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>