JVMCBC-1174 Native TLS library is missing
Motivation
==========
Fix a regression in core-io 2.2.7 that prevented using native TLS.

Modifications
=============
Specify architecture classifiers when unpacking the tcnative
native libraries. This ensures we get the x86_64 versions.

As a byproduct of this change, the `aarch_64` tcnative libraries
are no longer included in core-io. They had slipped in by mistake
and never worked because they weren't relocated.

Depend on platform-agnostic "*-classes-*" libraries instead of
"*-native-*" so we don't need to filter out the original
native libraries.

Change-Id: I725c97054f75b9782700a565003e0ea142021d86
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/183921
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>