Tests - workaround CoreTest issues
This test intermittently fails on CI as addService has been
called more times than expected.  After multiple rounds of debug
iteration it remains unclear exactly why this is, but it
appears the config intermittently gets emitted twice when
`configs.tryEmitNext(clusterConfig)` is called.

A Thread.sleep after the config emit reliably resolves the problem,
somehow.  Though this doesn't make much sense, since it would
seem that the config should be emitted twice or not regardless
of a sleep.

Since this seems to be a test-only issue, and since the functionality
this code is testing (config loading) is in the process of being
refactored and rewritten anyway, further root cause analysis does not
seem warranted at this time.  Especially as this failure periodically
but regularly breaks Gerrit validation.

Also leaving in-place the verbose logging that has been helping
investigate this problem, in case it aids future investigation.

Change-Id: I332f132f9d72fad0687f70205a20f5de45dcfee9
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/183967
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>