JVMCBC-1187: Core JVM SDK Protostellar support
Builds on the new CoreKvOps interface, adding initial support
for the Protostellar protocol.

The protocol is still in development so there are parts of this
code that will require later modification, particularly around
error handling, circuit breakers, retry handling, and diagnostics
& ping.

In addition only a handful of operations are currently implemented:
namely some of the KV ops, and queries.  The other operations
will be added over time.

Though a large change, it should have zero impact on existing
applications, as the modifications are largely segregated and will
only be used if "protostellar://" has been specified as the scheme.

Change-Id: I063f8a2c2042f2c322ba6d3f4844d2a9e8c065ac
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185134
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Graham Pople <graham.pople@couchbase.com>