JVMCBC-1210 Move mutateIn into core
Add `CoreKvOps.subdocMutate*` and a Classic implementation.

Modifications
-------------
Add an alternate SubdocMutateRequest constructor that accepts
the new CoreStoreSemantics and CoreSubdocMutateCommand,
and handles the detail of sorting commands (xattrs first).

Make SubdocMutateRequest's `bucketConfig` param nullable.
If null, the caller is responsible for checking bucket
capabilities.

Instead of fetching the bucket config first and doing
a pre-flight check to see if certain sub-document commands
are available, ClassicCoreKvOps.mutateIn always issues the
command, and does a post-hoc capability check to return
the appropriate FeatureNotAvailable exception only if the
request fails. This is simpler to implement and faster
on the happy path, at the cost of a less-efficient sad path.

Remove `MutateInResult.hashCode` and `equals`.
The implementations were flawed due to incorrect
comparisons, and of questionable value to begin with.

Change-Id: Iafd6ea4d39f3ec209e26e49ab6979a771ce10f1c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185289
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>