JVMCBC-1208: Add query support
Moves query behind a CoreQueryOps interface.

This allows the Classic and Protostellar implementations to be shared
easily by all JVM SDKs.

It's required finding abstractions for CoreQueryOptions, CoreQueryResult,
CoreQueryMetrics and so on.

It's an impactful change partly as transactions was using the low-level
QueryAccessor directly, as it needed to direct all queries in a
transaction to the same query node.  This targeting functionality
has been duplicated in CoreQueryOps, though by passing a
NodeIdentifier.  If this functionality is ever needed in Protostellar
(unlikely, since that protocol is explicitly against that use-case),
we will need to find a suitable abstraction.

The refactoring could probably go further.  The existing Classic
implementation has been kept mostly unchanged.  The existing
QueryResponse continues to be returned and gets wrapped into a
CoreQueryResult.  It would probably be possible to save some allocations
and push CoreQueryResult deeper into the query parsing engine.

Scala and Kotlin will need to be ported to CoreQueryOps in a
follow-up change.

Change-Id: I4859a26ba9244b3609e76fb84030f4f0188be8d3
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185470
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>