JVMCBC-1210 Restore pre-flight check for advanced subdoc features
Motivation
----------
The pre-flight check is a Chesterton's Fence.
If we decide it's safe to remove it, we can revert this
commit.

Modifications
-------------
Remove the post-hoc error translation.

Compose the "load the bucket config" future
with the "do the subdoc lookup" feature, and let
SubdocGetRequest validate the flags against the
bucket capabilities.

In BucketConfigUtil, reuse a single Duration
instead of creating a new one each time.

Change-Id: I109294c1b9d0d5b016c18ea145cf130073d45221
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185664
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>