SCBC-386: Support all KV ops
Converting Scala to use CoreKvOps for get*Replicas.

While largely straightforward, there is one unfortunate caveat.

There's an error with how the AsyncCollection.getAllReplicas interface
was originally designed.

Java has (the equivalent of) Future[Seq[Future[GetReplicaResult]]]
While Scala has              Seq[Future[GetReplicaResult]]

This wasn't too bad when the logic just had to get the number of replicas
to know what size of Seq to return.  Though it was still incorrect,
since getting the number of replicas requires the bucket config, which
can be a blocking operation.

But now, with the move to CoreKvOps (and more pertinently, Protostellar),
we don't know how many replicas there are - and hence what size Seq to
return - until all responses have been received.

So, despite this being an async non-blocking interface, we now have to block
the user's thread.

We will release note this change and advise users move to the reactive
implementation if they require non-blocking semantics.

Change-Id: I9670b8d7b77f3c2a3fcee198d7a5a1c0cf5c7393
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185696
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>