JVMCBC-1211: Add PreserveExpiry logic
Correcting the Protostellar expiry logic.

preserveExpiry and expiry are bundled in the same Protostellar
expiry fields with the following logic:

If preserveExpiry, don't send anything (it's the default).
Else, if convert the expiry (which could be 0 indicating to reset
the expiry, which is the SDK's default) and send that.

This new logic _should_ work, but other issues with STG are
preventing it being tested.

Change-Id: I9a657281e8efcd499fa2c4a8a3e2ce058678c867
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185703
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>