JVMCBC-1206 Remove SLF4J usage in non-Protostellar code
Motivation
----------
SLF4J is an optional dependency. Referencing it when it's
not present causes a class loading exception.

Modifications
-------------
When connection string has invalid port type,
throw an exception instead of logging a warning.

Remove unused SLF4J logger instances.

Remove the logging in CouchbaseThreadFactory
when a new thread is created.

Keep the SLF4J loggers in other Protostellar
code; we can address those uses later, and the
log output is useful in the meantime.

Change-Id: Ib951f8cea7219c825ea02c57090d694fad41a3bc
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185872
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <graham.pople@couchbase.com>