JVMCBC-1217: Move QueryIndexManager into core JCBC-2045: WatchQueryIndexOptions does not extend CommonOptions
This will prepare us better for upcoming changes to query_context.

This is a straightforward refactoring with no intentional functional
changes.

Changes have been made to option Built classes.  Technically these
are not marked @Internal.  However, they are only accessible via
the build() method, which is.

Fixing JCBC-2045 in passing.

Change-Id: I78960dcbd0b273113e197de7d61e235a4d8b9618
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/185963
Tested-by: Graham Pople <graham.pople@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>