Gardening - removing replaceTombstoneWithTxnXattrWithCASWhichHasChanged_pre71
This test has begun failing on 6.6.6: one of the operations used to
succeed, and now fails with DocumentExistsException.

Checking the comments for similar test replaceTombstoneWithTxnXattrWithCASWhichHasChanged_71Plus
reveals that DocumentExistsException is how things work as of 7.1.
Clearly the change has been included in what was backported to 6.6.6.

Since transactions is able to handle both behaviours without issue,
and because there's no easy way to make this test run on [6.6.6 - 7.1),
simply removing it.

Change-Id: Ib9c95b7e36a883e6e8a1c12bd85c10386f91c12c
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/186303
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>