JVMCBC-1230: Pathfind approach for scope-level FTS indexes
Adding a ScopeSearchIndexManager.  This is pathfinding and it's not
certain if this will be the final approach, so everything is at
@Volatile level.

Added in Core behind a CoreSearchIndexManager interface, so the
other SDKs can benefit and it will be easy to ultimately add
Protostellar support.

Regular SearchIndexManager could easily be refactored to use
this same CoreSearchIndexManager.  That work will be done
later when this approach is more settled.

Change-Id: Id8914e06d085a43f029c4543eed8a7798ea23014
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/187126
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>