JVMCBC-1197 Push DNS SRV lookup down into Core
Motivation
----------
DNS SRV lookup is not required for Protostellar
(at least, not the way we do lookups today).

Also, it's nice to always have a connection string,
even if the user gave us a set of `SeedNode`.

Modifications
-------------
Before this change, the language wrappers were responsible
for calling `seedNodesFromConnectionString`, which does the
DNS SRV lookup.

After this change, Core's constructor calls
`seedNodesFromConnectionString`.

As part of this change, Core no longer takes a set of `SeedNode`.
If the user passes a set of SeedNodes to `Cluster.connect`,
immediately convert it into an equivalent connection string,
so we always have a non-null connection string.

Change-Id: If280acae6d5e11af7866e92279caf91e539119a6
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/187165
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>