JVMCBC-1235: Move FTS into core
As a precursor to Protostellar support, moving the Java SDK's
FTS implementation into core-io, behind a CoreSearchOps
interface.

There should be no functional changes (beyond some initial
support for Protostellar messages).  This just as a
refactoring.

Removing SearchMock and the test that used it.  It's only
used by one test and would be an expensive fix.

Change-Id: I7ed893857bc9aa28d6dda22540a2f8de563fd962
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/187381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>