JVMCBC-1223: Fail fast and propagate AuthenticationFailureException if request fails due to bad credentials
Supports the ability to fast-fail operations iff the underlying
endpoint is failing to authenticate.

The mechanism that allows this is to add a new AUTHENTICATION_ERROR
RetryReason.  A custom RetryStrategy can then be used to handle
these and e.g. raise an AuthenticationFailureException rather than
the TimeoutException that would otherwise result.

Two sample such RetryStrategies are included in the test code,
and we could consider promoting them to be part of the SDK at a later
date.

To support this change, these changes are made:

* EndpointDiagnostics has a new AuthenticationStatus field, which is
  set at appropriate times during the SASL, SSL and endpoint connection
  paths.

* The two endpoint locators, which previously just checked if
  the config was being loaded and raised a RetryReason.BUCKET_OPEN_IN_PROGRESS
  (or similar) if so, now check to see if any of the KV GCCCP
  connections are in AuthenticationStatus.FAILED state.  And if so,
  raise a RetryReason.AUTHENTICATION_ERROR instead.
  See the comments in AuthErrorDecider for an explanation.

I haven't touched waitUntilReady in this changeset, as it's unclear
how that should work - since it currently does not use RetryStrategy,
and adding it could be considered a breaking change.

Change-Id: Ide9d9c7ff3350301eaa61ffb1cecf9cf609d341d
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/188096
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>