KCBC-118 LookupIn "exists" returns false when it should throw exception
Motivation
----------
LookupIn "exists" erroneously returns false when the document
is too deep for the server to parse it, and also when the user
does not have permission to access the target XATTR.

Don't want to mask the actual error condition, or mislead the user
about the existence of a field, when in fact we can't tell if
the field exists.

Modifications
-------------
When processing the "exists" response, instead of looking for
the errors to propagate, look for the errors that mean
the field definitely does not exist. This is a shorter list,
and less likely to change if new error codes are added later.

Change-Id: I35bbd7c77e0ae298e26cfa91b04d9ceef6229ee4
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/188574
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>