JCBC-1982 - Cancel Range Scans when Reactor chain is cancelled.
For sampling scans, when the limit is reached by the take(limit)
operator, the reactor chains are cancelled. This would leave
any rangescans alive on the server until the server decides they have
been abandoned and removes them.

To allow the RangeScanOrchestrator to cancel them explicitly,
the code that creates and uses a rangescan is protected from cancels
with Reactor.shieldFromCancellation. This isolates the shielded
flux from cancels, such that it can send a RangeScanCancel request if
necessary. There is also a needToCancel flag that is set
outside the shielded flux to indicate that the rangescan
should be cancelled, return no more items, and not be repeated.

For testing, a sample scan that reaches its limit before a
rangescan is COMPLETE will need to be cancelled. Such a
sample scan is repeated many times to ensure that there
are not leaked rangescans which would run the server out of
resources.

Change-Id: I40e20e3205c700a9d0b9eeedb1319386da6d72f4
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/188697
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>