JVMCBC-1223: Fail fast and propagate AuthenticationFailureException if request fails due to bad credentials
Reducing the scope of the (unreleased) previous fix.

This previous fix included adding an AuthenticationStatus field
into EndpointDiagnostics.  Tests checked that it was in expected
state for all services, which passed locally, but on CI uncovered
bugs with this approach.  Including that the MANAGER service
is incorrectly assuming successful authentication as soon as its
endpoint is connected, rather than on a non-401 response.  Similar
bugs likely exist for the other services.

Providing a correct, fully tested working public AuthenticationStatus
for all services is a much larger requirement than is needed for the
the ticket.  For this, we just need to internally be able to check
the GCCCP KV AuthenticationStatus.  So, this has been moved into
a new EndpointDiagnosticsInternal to reduce the scope of the change.

And, removing the bug where the AuthenticationStatus is set to SUCCESS
after any endpoint connects.  it's now only set to SUCCESS when
the SASL authenticator succeeds.

And, removing AuthenticationStatus completely from ProtostellarEndpoint
until I'm more certain it makes sense with Protostellar's stateless
connection-less model.

Change-Id: I64f2811e3445d5d48c7404c8e1db328b528f6937
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/188822
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>