JVMCBC-1277 KV Range Scan: Represent scan term as String instead of a byte array
Motivation
----------
Protostellar wants the term to be a valid UTF-8 string.

Modifications
-------------
Remove ScanTerm constructors that take byte array.

Use a valid unicode character to represent the max scan term,
instead of 0xff which is not valid UTF-8.

Update FIT executors to throw UnsupportedOperation when
performed tries to specify scan term as a byte array.

Change-Id: I6c8b30ef602f96adc98729837a915b577fffffee
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/189958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>