JVMCBC-1281 FTS: Wrong sort order when mixing SearchSort and String
Motivation
----------
A regression in Java SDK 3.4.5 and Scala SDK 1.4.5
( commit 2ab79b64ab) could cause Full-Text Search
results to be sorted in the wrong order if the
sort option was specified using a mix of strings
and `SearchSort` objects.

Modifications
-------------
Add a CoreSearchSortString class to represent
a sort tier specified as a string.

Store all sort tiers in the same list. This
preserves the order among tiers.

Removed CoreSearchOptions.sortStrings().

Add CoreSearchSort.toJsonNode, and use it to
build the JSON representation of the sort tier.
(injectParams() doesn't work for string sorts,
because they are represented as JSON Strings
instead of Objects.)

Change-Id: I090a20a0cba1db63694ee42957c0740daad2a2e0
Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/190146
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>